projeto-lucas/requirements.txt
Flask
python-dotenv
requests
pymongo
werkzeug.utils
werkzeug.security